Перенес инициализацию стора из мидлвара в контроллер
