+++ src/server/index.d.ts
@@ -1,6 +1,3 @@
-/* eslint-disable import/no-unresolved */
-import { store } from 'client';
-
 declare global {
     namespace Express {
         interface Request {
@@ -8,12 +5,5 @@ declare global {
             logger: () => void;
         }
 
-        interface Response {
-
-            store: ReturnType<typeof store>,
-
-            renderApp(): void
-
-        }
     }
 }
